Remove wit-bindgen generated code
Now they will be generated under src_managed directory
