--- examples/helloworld-wasi/src/main/scala/helloworld/wasi/io/error.scala
@@ -1,44 +1,0 @@
-package helloworld.wasi.io
-
-package object error {
-
-  // Resources
-  /** A resource which represents some error information.
-   *
-   *  The only method provided by this resource is `to-debug-string`,
-   *  which provides some human-readable information about the error.
-   *
-   *  In the `wasi:io` package, this resource is returned through the
-   *  `wasi:io/streams/stream-error` type.
-   *
-   *  To provide more specific error information, other interfaces may
-   *  provide functions to further "downcast" this error into more specific
-   *  error information. For example, `error`s returned in streams derived
-   *  from filesystem types to be described using the filesystem's own
-   *  error-code type, using the function
-   *  `wasi:filesystem/types/filesystem-error-code`, which takes a parameter
-   *  `borrow<error>` and returns
-   *  `option<wasi:filesystem/types/error-code>`.
-   *
-   *  The set of functions which can "downcast" an `error` into a more
-   *  concrete type is open.
-   */
-  @scala.scalajs.wit.annotation.WitResourceImport("wasi:io/error@0.2.0", "error")
-  trait Error {
-    /** Returns a string that is suitable to assist humans in debugging
-     *  this error.
-     *
-     *  WARNING: The returned string should not be consumed mechanically!
-     *  It may change across platforms, hosts, or other implementation
-     *  details. Parsing this string is a major platform-compatibility
-     *  hazard.
-     */
-    @scala.scalajs.wit.annotation.WitResourceMethod("[method]error.to-debug-string")
-    def methodErrorToDebugString(self: Error): String = scala.scalajs.wit.native
-    @scala.scalajs.wit.annotation.WitResourceDrop
-    def close(): Unit = scala.scalajs.wit.native
-  }
-  object Error {
-  }
-
-}
